requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic-settings==2.6.1
anthropic==0.39.0
higgsfield-client==0.1.0
google-cloud-texttospeech==2.17.2
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httpx==0.28.0
beautifulsoup4==4.12.3
pymupdf==1.24.13
apscheduler==3.10.4
sse-starlette==2.1.3
aiofiles==24.1.0
python-dotenv==1.0.1
pillow==11.0.0
